DataTable reutilizada para COEQ e COAPS funcionando!
