+++ src/features/acf/frontend/diabetes/modules/AcfPage/modules/PanelSelector/modules/List/modules/DataTable/modules/common/DataTable/modules/EmptyTableMessage.tsx
@@ -1,4 +1,3 @@
-//TODO: Remover este arquivo depois que a DataTable genérica estiver pronta
 export const EmptyTableMessage: React.FC = () => {
     return (
         <p
